+++ demos/laminate/src/main.rs
@@ -11,30 +11,30 @@ use polygon_unionfind::{
 };
 use undoredo::UndoRedo;
 
-type DemoPolygon = PolygonWithData<i64, ()>;
-type DemoNegated = RecordingNegated<i64, DemoPolygon>;
+type DemoPolygon = PolygonWithData<i32, ()>;
+type DemoNegated = RecordingNegated<i32, DemoPolygon>;
 type DemoInnerLayer = Paralleled<DemoNegated>;
 type DemoLayer = Paralleled<DemoInnerLayer>;
-type DemoTransition = Paralleled<RecordingPolygonSet<i64, DemoPolygon>>;
-type DemoLaminate = RecordingLaminate<i64, DemoPolygon>;
+type DemoTransition = Paralleled<RecordingPolygonSet<i32, DemoPolygon>>;
+type DemoLaminate = RecordingLaminate<i32, DemoPolygon>;
 type DemoLayersDelta = LaminateDelta<
-    i64,
+    i32,
     DemoPolygon,
     Paralleled<
         Paralleled<
             Negated<
-                PolygonSetHalfDelta<i64, DemoPolygon>,
+                PolygonSetHalfDelta<i32, DemoPolygon>,
                 Inflated<
-                    PolygonUnionFindHalfDelta<i64, DemoPolygon>,
-                    core::marker::PhantomData<i64>,
+                    PolygonUnionFindHalfDelta<i32, DemoPolygon>,
+                    core::marker::PhantomData<i32>,
                 >,
             >,
         >,
     >,
-    Paralleled<PolygonSetHalfDelta<i64, DemoPolygon>>,
+    Paralleled<PolygonSetHalfDelta<i32, DemoPolygon>>,
 >;
-const PRIMARY_INFLATION: i64 = 0;
-const PARALLEL_INFLATION: i64 = 50;
+const PRIMARY_INFLATION: i32 = 0;
+const PARALLEL_INFLATION: i32 = 50;
 
 fn frame_polygon() -> DemoPolygon {
     DemoPolygon {
@@ -44,12 +44,12 @@ fn frame_polygon() -> DemoPolygon {
     }
 }
 
-fn new_negated_layer(offset: i64) -> DemoNegated {
-    let mut minuend: RecordingPolygonSet<i64, DemoPolygon> = RecordingPolygonSet::new();
+fn new_negated_layer(offset: i32) -> DemoNegated {
+    let mut minuend: RecordingPolygonSet<i32, DemoPolygon> = RecordingPolygonSet::new();
     let _ = minuend.add(frame_polygon());
     RecordingNegated::new(
         minuend,
-        RecordingInflated::<i64, DemoPolygon>::new(offset),
+        RecordingInflated::<i32, DemoPolygon>::new(offset),
     )
 }
 
@@ -84,9 +84,9 @@ fn convex_hull(points: &[[i32; 2]]) -> Vec<[i32; 2]> {
     let mut pts: Vec<[i32; 2]> = points.to_vec();
     pts.sort_by(|a, b| a[0].cmp(&b[0]).then_with(|| a[1].cmp(&b[1])));
 
-    fn cross(o: [i32; 2], a: [i32; 2], b: [i32; 2]) -> i64 {
-        (a[0] as i64 - o[0] as i64) * (b[1] as i64 - o[1] as i64)
-            - (a[1] as i64 - o[1] as i64) * (b[0] as i64 - o[0] as i64)
+    fn cross(o: [i32; 2], a: [i32; 2], b: [i32; 2]) -> i32 {
+        (a[0] as i32 - o[0] as i32) * (b[1] as i32 - o[1] as i32)
+            - (a[1] as i32 - o[1] as i32) * (b[0] as i32 - o[0] as i32)
     }
 
     let mut lower = Vec::new();
@@ -136,7 +136,7 @@ fn polygon_from_ring_i32(ring: Vec<[i32; 2]>) -> DemoPolygon {
     DemoPolygon {
         exterior: ring
             .into_iter()
-            .map(|[x, y]| [i64::from(x), i64::from(y)])
+            .map(|[x, y]| [i32::from(x), i32::from(y)])
             .collect(),
         interiors: vec![],
         data: (),
@@ -156,13 +156,13 @@ fn random_polygon_at_screen_click(center: Vec2, zoom: f32, mx: f32, my: f32) -> 
 }
 
 fn draw_polygon_set_lines(
-    polygon_set: &RecordingPolygonSet<i64, DemoPolygon>,
+    polygon_set: &RecordingPolygonSet<i32, DemoPolygon>,
     center: Vec2,
     zoom: f32,
     color: Color,
 ) {
     for (_index, polygon) in polygon_set.polygons().as_ref().iter() {
-        let rings: Vec<&[[i64; 2]]> = std::iter::once(polygon.exterior.as_slice())
+        let rings: Vec<&[[i32; 2]]> = std::iter::once(polygon.exterior.as_slice())
             .chain(polygon.interiors.iter().map(Vec::as_slice))
             .collect();
         for ring in rings.iter().copied() {
